+++ server/domain/member.go
@@ -1,0 +1,13 @@
+package domain
+
+type Member struct {
+	Id       int
+	Name     string
+	Comment  string
+	Lang     string
+	Github   string
+	Twitter  string
+	Question string
+	Room     string
+}
+
