+++ server/domain/room.go
@@ -1,0 +1,13 @@
+package domain
+
+import (
+	"time"
+)
+
+type Room struct {
+	Id        string
+	Name      string
+	MaxCount  int    // roomの最大人数
+	Status    string // roomのstatus
+	CreatedAt time.Time
+}
