+++ server/main.go
@@ -3,13 +3,11 @@ package main
 import (
 	"fmt"
 	"os"
-
-	"github.com/Doer-org/hack-camp_vol5_2022/server/controller"
 )
 
 func main() {
 
-	router := controller.InitRouter()
+	router := InitRouter()
 
 	port := fmt.Sprintf(":%s", os.Getenv("PORT"))
 	router.Run(port)
