+++ server/repository/room.go
@@ -1,0 +1,42 @@
+package repository
+
+import (
+	"github.com/Doer-org/hack-camp_vol5_2022/server/db"
+	"github.com/Doer-org/hack-camp_vol5_2022/server/domain"
+)
+
+func GetAllRoom() (rooms []domain.Room) {
+	db := db.NewDB()
+	defer db.Conn.Close()
+
+	db.Conn.Find(&rooms)
+	return
+}
+
+func NewRoom(room domain.Room) domain.Room {
+	db := db.NewDB()
+	defer db.Conn.Close()
+
+	db.Conn.Save(&room)
+	return room
+}
+
+func GetRoomByID(id string) domain.Room {
+	db := db.NewDB()
+	defer db.Conn.Close()
+
+	room := domain.Room{Id: id}
+	// Idでroomを取得し,ほかのカラムも情報も更新
+	db.Conn.First(&room)
+	return room
+}
+
+func ChangeRoomStatus(id string) domain.Room {
+	db := db.NewDB()
+	defer db.Conn.Close()
+
+	room := domain.Room{Id: id}
+	room.Status = "finished"
+	db.Conn.Save(&room)
+	return room
+}
